+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Default model usage wasn't working properly
+
 ### Security
 
 ## [v1.2.1] - 2025-11-05